pyproject.toml
[project]
name = "circlefund-bank"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "apscheduler>=3.11.0",
    "bcrypt==3.2.2",
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
